+++ packages/conventional-commits/commitizen-config/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.0.4 (2020-04-15)
+
+
+* extract conventional-commits packages ([5eba84f](https://github.com/hzdg/hz-core/commit/5eba84f))
+* publish ([aef97cd](https://github.com/hzdg/hz-core/commit/aef97cd))
+* publish ([8e76b89](https://github.com/hzdg/hz-core/commit/8e76b89))
+* publish ([6b78b04](https://github.com/hzdg/hz-core/commit/6b78b04))
+* namespace under @hzdg ([b1b92f3](https://github.com/hzdg/hz-core/commit/b1b92f3))
+
+### @hzcore/commitizen-config
+
+* our new lintconfig is ok with process.exit ([00474ec](https://github.com/hzdg/hz-core/commit/00474ec))
+
+
 ## [1.0.3](https://github.com/hzdg/hz-core/compare/@hzcore/commitizen-config@1.0.2...@hzcore/commitizen-config@1.0.3) (2019-05-31)
 
 **Note:** Version bump only for package @hzcore/commitizen-config