+++ packages/conventional-commits/conventional-changelog/CHANGELOG.md
@@ -3,6 +3,21 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.2.0 (2020-04-15)
+
+
+* extract conventional-commits packages ([5eba84f](https://github.com/hzdg/hz-core/commit/5eba84f))
+* publish ([aef97cd](https://github.com/hzdg/hz-core/commit/aef97cd))
+* publish ([6b78b04](https://github.com/hzdg/hz-core/commit/6b78b04))
+* namespace under @hzdg ([b1b92f3](https://github.com/hzdg/hz-core/commit/b1b92f3))
+* conventional-changelog ([4a4a298](https://github.com/hzdg/hz-core/commit/4a4a298))
+
+### @hzcore/conventional-changelog
+
+* support breaking change notes ([b8a274e](https://github.com/hzdg/hz-core/commit/b8a274e))
+* changelog templating ([5527ffd](https://github.com/hzdg/hz-core/commit/5527ffd))
+
+
 # [1.1.0](https://github.com/hzdg/hz-core/compare/@hzcore/conventional-changelog@1.0.1...@hzcore/conventional-changelog@1.1.0) (2019-05-31)