+++ packages/utilities/create-package/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.1.0](https://github.com/hzdg/hz-core/compare/create-package@1.0.5...create-package@1.1.0) (2020-04-15)
+
+
+* remote @ts-ignores; add missing types ([b7bf48c](https://github.com/hzdg/hz-core/commit/b7bf48c))
+* namespace under @hzdg ([b1b92f3](https://github.com/hzdg/hz-core/commit/b1b92f3))
+* strip hzcore- prefix from package directories ([6aafd85](https://github.com/hzdg/hz-core/commit/6aafd85))
+
+### create-package
+
+* default to use `use-this-hook` ([6b6b7e3](https://github.com/hzdg/hz-core/commit/6b6b7e3))
+
+
 ## [1.0.5](https://github.com/hzdg/hz-core/compare/create-package@1.0.4...create-package@1.0.5) (2020-03-12)